+++ README.md
@@ -24,7 +24,13 @@
 
 ## How to Use
 
-Simply download the `index.html` file and open it in a modern web browser. No installation or build steps are required.
+Simply download the repository (or the `index.html` with the `src` folder) and open `index.html` in a modern web browser. All modules are loaded directly without a bundler.
+
+## Project Structure
+
+- `src/core/` – neural network implementation (`network.js`)
+- `src/ui/` – DOM setup and event handling (`main.js`)
+- `src/visualization/` – canvas drawing utilities
 
 ## Technology
