+++ package.json
@@ -1,0 +1,7 @@
+{
+  "name": "oblix",
+  "type": "module",
+  "scripts": {
+    "test": "node tests/run-tests.js"
+  }
+}
